backend/requirements.txt
motor
python-decouple
pydantic[email]
typing-extensions
fastapi
passlib[bcrypt]
PyJWT
uvicorn[standard]
fastapi-mail
python-multipart

